@@ -9,7 +9,7 @@
     using Polly.Retry;
     using System;
     using System.Collections.Generic;
-    using System.Data.SqlClient;
+    using Microsoft.Data.SqlClient;
     using System.Globalization;
     using System.IO;
     using System.IO.Compression;
