@@ -6,7 +6,7 @@ using Microsoft.Extensions.Options;
 using Ordering.BackgroundTasks.Events;
 using System;
 using System.Collections.Generic;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using System.Threading;
 using System.Threading.Tasks;
 
