@@ -1,9 +1,10 @@
-﻿import { Component, OnInit }    from '@angular/core';
+﻿
+import {throwError as observableThrowError,  Observable } from 'rxjs';
+import { Component, OnInit }    from '@angular/core';
 import { CampaignsService }        from './campaigns.service';
 import { ICampaign }               from '../shared/models/campaign.model';
 import { IPager }               from '../shared/models/pager.model';
 import { ConfigurationService } from '../shared/services/configuration.service';
-import { Observable } from 'rxjs';
 import { catchError } from 'rxjs/operators';
 
 @Component({
@@ -61,7 +62,7 @@ export class CampaignsComponent implements OnInit {
 
     private handleError(error: any) {
         this.errorReceived = true;
-        return Observable.throw(error);
+        return observableThrowError(error);
     }  
 }
 
