@@ -1,5 +1,6 @@
+
+import {throwError as observableThrowError,  Observable } from 'rxjs';
 import { Component, OnInit } from '@angular/core';
-import { Observable } from 'rxjs';
 import { catchError } from 'rxjs/operators';
 
 import { OrdersService } from '../orders.service';
@@ -90,7 +91,7 @@ export class OrdersNewComponent implements OnInit {
             .pipe(catchError((errMessage) => {
                 this.errorReceived = true;
                 this.isOrderProcessing = false;
-                return Observable.throw(errMessage); 
+                return observableThrowError(errMessage); 
             }))
             .subscribe(res => {
                 this.router.navigate(['orders']);
